+++ DelvUI/Helpers/Utils.cs
@@ -359,10 +359,11 @@ namespace DelvUI.Helpers
             };
         }
 
-        public static Status HasTankInvulnerability(BattleChara actor)
+        public static Status? GetTankInvulnerabilityID(BattleChara actor)
         {
             Status? tankInvulnBuff = actor.StatusList.FirstOrDefault(o => o.StatusId is 810 or 811 or 1302 or 409 or 1836 or 82);
-            return tankInvulnBuff!;
+
+            return tankInvulnBuff;
         }
 
         public static bool IsOnCleanseJob()